Fixed block falling, and made them delete at the bottom.
This required switching the moving test system from being independent of the gameobjects, to referencing the gameobjects. The problem was that if you tried to get the position of the current object in map, it would give an incorrect result because it was moving, hence you got null errors.
Another change was the rotation was disabled in animation until more time is put into fixing it.
There were also fixes to the move methods, where they would incorrectly move because they were not being called in the right order.